requirements.txt
discord.py
requests
pyyaml
sqlalchemy[asyncio]
aiosqlite
